s-skeinforge: first commit.
This is a convenience script I've written for myself. It is an attempt
at addressing the usability mess that skeinforge is, and at the same
time an attempt at making skeinforge somewhat more scriptable.

The script isn't developed in conjunction with skeinforge, and may as
such not work with your version. The way it interacts with skeinforge's
settings repository (that of yanking over its .csv's and constructing a
-p PREFRENCESDIRECTORY-worthy working copy) is a reasonably stable way
to do it, though---currently, it's working with my skeinforge
2013.01.28-1, as found in Debian Sid.